Fixed bugs after debugging
Full Changelog:
-Changed SortedArray(IComparable<T> comparable, int size,
ISortingAlgorithm<T> algorithm, T... elements) and
SortedArray(IComparable<T> comparable, int size, ESortingAlgorithm
algorithm, T... elements) to not use this(int maxSize, , IComparable<T>
comparable), since this constructor also allocates a new array, 
which is obsolete
-Changed SortedArray(IComparable<T> comparable, ESortingAlgorithm
algorithm, T... elements)'s second parameter name from "algo" to
"algorithm".
-Renamed the method delete(int index) to deleteAtIndex(int index). In
the case that the generic parameter is of the type Integer, Java would
not know which to choose, so in the case that the User wanted to delete
the element "10", Java would try to delete the element at the index 10.
-In order of the previous change, the call of delete(int index) in
Iterator.remove() has been altered to deleteAtIndex(...).
-In equals(...), removed the check for IComparable equality, since these
only differ in different implementations and it is not possible to test
these on equality. Also updated the doc accordingly.
-In equals(...), fixed the checking if the elements are the same.
(Previously, the algorithm would iterate over this.elements and check
each element is equal to the Object passed to the method)
-In equals(...), a check if the Object passed to the method is an
instance of SortedArray.
-In equals(...), a check if the Object passed to the method is not null.
-In setSize(...), the amount of objects copied is now limited to
'currentSize', if 'newSize' < currentSize,  'newSize' is the amount of
objects copied, otherwise it is 'currentSize'.
-setSize(...) now updates 'currentSize' if the array has been shrinked. 
-setSize(...) now checks, if 'newSize' is unequal the maximumSize.