Expanded RuntimeTester functionality
Added a variety of features:
-Added a init phase -> a runnable that is called each time before the
runnable with the code to test is called, but does not contribute to the
final results of the test.
-Switched from java.lang.Runnable to a custom Runnable that holds a
reference to the calling RuntimeTester.
-The usage of the change from the first point is, that the RuntimeTester
holds an Object[] that is meant to be filled by the user during the init
phase (with data that is needed during the test).